@@ -1,6 +1,6 @@
 import ntpath
 import os
-from typing import Dict, List, Optional, Union
+from typing import Optional, Union
 
 import requests
 
@@ -17,12 +17,14 @@ from ytmusicapi.parsers.library import (
 from ytmusicapi.parsers.uploads import parse_uploaded_items
 
 from ..auth.types import AuthType
+from ..enums import ResponseStatus
+from ..exceptions import YTMusicUserError
 from ._protocol import MixinProtocol
 from ._utils import prepare_order_params, validate_order_parameter
 
 
 class UploadsMixin(MixinProtocol):
-    def get_library_upload_songs(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[Dict]:
+    def get_library_upload_songs(self, limit: Optional[int] = 25, order: Optional[str] = None) -> list[dict]:
         """
         Returns a list of uploaded songs
 
@@ -69,7 +71,7 @@ class UploadsMixin(MixinProtocol):
 
         return songs
 
-    def get_library_upload_albums(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[Dict]:
+    def get_library_upload_albums(self, limit: Optional[int] = 25, order: Optional[str] = None) -> list[dict]:
         """
         Gets the albums of uploaded songs in the user's library.
 
@@ -90,7 +92,7 @@ class UploadsMixin(MixinProtocol):
 
     def get_library_upload_artists(
         self, limit: Optional[int] = 25, order: Optional[str] = None
-    ) -> List[Dict]:
+    ) -> list[dict]:
         """
         Gets the artists of uploaded songs in the user's library.
 
@@ -109,7 +111,7 @@ class UploadsMixin(MixinProtocol):
             response, lambda additionalParams: self._send_request(endpoint, body, additionalParams), limit
         )
 
-    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> List[Dict]:
+    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> list[dict]:
         """
         Returns a list of uploaded tracks for the artist.
 
@@ -158,7 +160,7 @@ class UploadsMixin(MixinProtocol):
 
         return items
 
-    def get_library_upload_album(self, browseId: str) -> Dict:
+    def get_library_upload_album(self, browseId: str) -> dict:
         """
         Get information and tracks of an album associated with uploaded tracks
 
@@ -200,7 +202,7 @@ class UploadsMixin(MixinProtocol):
         album["duration_seconds"] = sum_total_duration(album)
         return album
 
-    def upload_song(self, filepath: str) -> Union[str, requests.Response]:
+    def upload_song(self, filepath: str) -> Union[ResponseStatus, requests.Response]:
         """
         Uploads a song to YouTube Music
 
@@ -209,21 +211,19 @@ class UploadsMixin(MixinProtocol):
         """
         self._check_auth()
         if not self.auth_type == AuthType.BROWSER:
-            raise Exception("Please provide browser authentication before using this function")
+            raise YTMusicUserError("Please provide browser authentication before using this function")
         if not os.path.isfile(filepath):
-            raise Exception("The provided file does not exist.")
+            raise YTMusicUserError("The provided file does not exist.")
 
         supported_filetypes = ["mp3", "m4a", "wma", "flac", "ogg"]
         if os.path.splitext(filepath)[1][1:] not in supported_filetypes:
-            raise Exception(
+            raise YTMusicUserError(
                 "The provided file type is not supported by YouTube Music. Supported file types are "
                 + ", ".join(supported_filetypes)
             )
 
         headers = self.headers.copy()
-        upload_url = (
-            "https://upload.youtube.com/upload/usermusic/http?authuser=%s" % headers["x-goog-authuser"]
-        )
+        upload_url = f"https://upload.youtube.com/upload/usermusic/http?authuser={headers['x-goog-authuser']}"
         filesize = os.path.getsize(filepath)
         body = ("filename=" + ntpath.basename(filepath)).encode("utf-8")
         headers.pop("content-encoding", None)
@@ -239,11 +239,11 @@ class UploadsMixin(MixinProtocol):
             response = requests.post(upload_url, data=file, headers=headers, proxies=self.proxies)
 
         if response.status_code == 200:
-            return "STATUS_SUCCEEDED"
+            return ResponseStatus.SUCCEEDED
         else:
             return response
 
-    def delete_upload_entity(self, entityId: str) -> Union[str, Dict]:  # pragma: no cover
+    def delete_upload_entity(self, entityId: str) -> Union[str, dict]:  # pragma: no cover
         """
         Deletes a previously uploaded song or album
 
@@ -260,6 +260,6 @@ class UploadsMixin(MixinProtocol):
         response = self._send_request(endpoint, body)
 
         if "error" not in response:
-            return "STATUS_SUCCEEDED"
+            return ResponseStatus.SUCCEEDED
         else:
             return response["error"]
