@@ -1,23 +1,28 @@
-from typing import List, Optional
+from typing import Optional
 
 from ..helpers import to_int
 from .songs import *
 
 
-def parse_playlist_header(response: Dict) -> Dict[str, Any]:
-    playlist: Dict[str, Any] = {}
-    own_playlist = "musicEditablePlaylistDetailHeaderRenderer" in response["header"]
-    if not own_playlist:
-        header = response["header"]["musicDetailHeaderRenderer"]
-        playlist["privacy"] = "PUBLIC"
+def parse_playlist_header(response: dict) -> dict[str, Any]:
+    playlist: dict[str, Any] = {}
+    editable_header = nav(response, [*HEADER, *EDITABLE_PLAYLIST_DETAIL_HEADER], True)
+    playlist["owned"] = editable_header is not None
+    playlist["privacy"] = "PUBLIC"
+    if editable_header is not None:  # owned playlist
+        header = nav(editable_header, HEADER_DETAIL)
+        playlist["privacy"] = editable_header["editHeader"]["musicPlaylistEditHeaderRenderer"]["privacy"]
     else:
-        header = response["header"]["musicEditablePlaylistDetailHeaderRenderer"]
-        playlist["privacy"] = header["editHeader"]["musicPlaylistEditHeaderRenderer"]["privacy"]
-        header = header["header"]["musicDetailHeaderRenderer"]
-    playlist["owned"] = own_playlist
+        header = nav(response, HEADER_DETAIL, True)
+        if header is None:
+            header = nav(
+                response, [*TWO_COLUMN_RENDERER, *TAB_CONTENT, *SECTION_LIST_ITEM, *RESPONSIVE_HEADER]
+            )
 
     playlist["title"] = nav(header, TITLE_TEXT)
-    playlist["thumbnails"] = nav(header, THUMBNAIL_CROPPED)
+    playlist["thumbnails"] = nav(header, THUMBNAIL_CROPPED, True)
+    if playlist["thumbnails"] is None:
+        playlist["thumbnails"] = nav(header, THUMBNAILS)
     playlist["description"] = nav(header, DESCRIPTION, True)
     run_count = len(nav(header, SUBTITLE_RUNS))
     if run_count > 1:
@@ -39,14 +44,16 @@ def parse_playlist_header(response: Dict) -> Dict[str, Any]:
         playlist["duration"] = (
             None if not has_duration else second_subtitle_runs[has_views + has_duration]["text"]
         )
-        song_count = second_subtitle_runs[has_views + 0]["text"].split(" ")
-        song_count = to_int(song_count[0]) if len(song_count) > 1 else 0
+        song_count_text = second_subtitle_runs[has_views + 0]["text"]
+        song_count_search = re.search(r"\d+", song_count_text)
+        # extract the digits from the text, return 0 if no match
+        song_count = to_int(song_count_search.group()) if song_count_search is not None else 0
         playlist["trackCount"] = song_count
 
     return playlist
 
 
-def parse_playlist_items(results, menu_entries: Optional[List[List]] = None, is_album=False):
+def parse_playlist_items(results, menu_entries: Optional[list[list]] = None, is_album=False):
     songs = []
     for result in results:
         if MRLIR not in result:
@@ -60,8 +67,8 @@ def parse_playlist_items(results, menu_entries: Optional[List[List]] = None, is_
 
 
 def parse_playlist_item(
-    data: Dict, menu_entries: Optional[List[List]] = None, is_album=False
-) -> Optional[Dict]:
+    data: dict, menu_entries: Optional[list[list]] = None, is_album=False
+) -> Optional[dict]:
     videoId = setVideoId = None
     like = None
     feedback_tokens = None
@@ -90,17 +97,64 @@ def parse_playlist_item(
             if "menu" in data:
                 like = nav(data, MENU_LIKE_STATUS, True)
 
-    title = get_item_text(data, 0)
+    isAvailable = True
+    if "musicItemRendererDisplayPolicy" in data:
+        isAvailable = data["musicItemRendererDisplayPolicy"] != "MUSIC_ITEM_RENDERER_DISPLAY_POLICY_GREY_OUT"
+
+    # For unavailable items indexes are preset,
+    # because meaning of the flex column cannot be found using navigationEndpoint
+    title_index = 0 if isAvailable is False else None
+    artist_index = 1 if isAvailable is False else None
+    album_index = 2 if isAvailable is False else None
+    user_channel_indexes = []
+    unrecognized_index = None
+
+    for index in range(len(data["flexColumns"])):
+        flex_column_item = get_flex_column_item(data, index)
+        navigation_endpoint = nav(flex_column_item, [*TEXT_RUN, "navigationEndpoint"], True)
+
+        if not navigation_endpoint:
+            unrecognized_index = index if unrecognized_index is None else unrecognized_index
+            continue
+
+        if "watchEndpoint" in navigation_endpoint:
+            title_index = index
+        elif "browseEndpoint" in navigation_endpoint:
+            page_type = nav(
+                navigation_endpoint,
+                [
+                    "browseEndpoint",
+                    "browseEndpointContextSupportedConfigs",
+                    "browseEndpointContextMusicConfig",
+                    "pageType",
+                ],
+            )
+
+            # MUSIC_PAGE_TYPE_ARTIST for regular songs, MUSIC_PAGE_TYPE_UNKNOWN for uploads
+            if page_type == "MUSIC_PAGE_TYPE_ARTIST" or page_type == "MUSIC_PAGE_TYPE_UNKNOWN":
+                artist_index = index
+            elif page_type == "MUSIC_PAGE_TYPE_ALBUM":
+                album_index = index
+            elif page_type == "MUSIC_PAGE_TYPE_USER_CHANNEL":
+                user_channel_indexes.append(index)
+
+    # Extra check for rare songs, where artist is non-clickable and does not have navigationEndpoint
+    if artist_index is None and unrecognized_index is not None:
+        artist_index = unrecognized_index
+
+    # Extra check for non-song videos, last channel is treated as artist
+    if artist_index is None and user_channel_indexes:
+        artist_index = user_channel_indexes[-1]
+
+    title = get_item_text(data, title_index) if title_index is not None else None
     if title == "Song deleted":
         return None
 
-    flex_column_count = len(data["flexColumns"])
+    artists = parse_song_artists(data, artist_index) if artist_index is not None else None
 
-    artists = parse_song_artists(data, 1)
+    album = parse_song_album(data, album_index) if album_index is not None else None
 
-    album = parse_song_album(data, flex_column_count - 1) if not is_album else None
-
-    views = get_item_text(data, 2) if flex_column_count == 4 or is_album else None
+    views = get_item_text(data, 2) if is_album else None
 
     duration = None
     if "fixedColumns" in data:
@@ -110,10 +164,6 @@ def parse_playlist_item(
             duration = get_fixed_column_item(data, 0)["text"]["runs"][0]["text"]
 
     thumbnails = nav(data, THUMBNAILS, True)
-
-    isAvailable = True
-    if "musicItemRendererDisplayPolicy" in data:
-        isAvailable = data["musicItemRendererDisplayPolicy"] != "MUSIC_ITEM_RENDERER_DISPLAY_POLICY_GREY_OUT"
 
     isExplicit = nav(data, BADGE_LABEL, True) is not None
 
